+++ demos/react-demo/src/Player.js
@@ -38,8 +38,8 @@ const Player = () => {
           // https://open.ys7.com/help/1772?h=domain
           // domain默认是 https://open.ys7.com, 如果是私有化部署或海外的环境，请配置对应的domain
           // The default domain is https://open.ys7.com If it is a private deployment or overseas (outside of China) environment, please configure the corresponding domain
-          domain: "https://open.ys7.com"
-        }
+          domain: "https://open.ys7.com",
+        },
       });
     }
   }, []);
@@ -132,7 +132,9 @@ const Player = () => {
   return (
     <div>
       <h2>视频模式使用示例：</h2>
-      <div id="player-container"></div>
+      <div>
+        <div id="player-container"></div>
+      </div>
       <div>
         url: <input type="text" ref={urlRef} style={{ width: 500 }} />
       </div>